requirement.txt
flask
whisper
ffmpeg-python